Refactor validators: reorganize in right files
